add shared arena hashmap
